refactor(config): extract runtime configuration module
